website/requirements.txt
Django==5.2.5
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-unfold==0.20.0
Pillow>=10.0.0
python-decouple==3.8
requests>=2.31.0
